+++ crate/rt/src/commands.rs
@@ -1,10 +1,11 @@
 pub use self::{
-    diff_cmd::DiffCmd, ensure_cmd::EnsureCmd, state_desired_discover_cmd::StateDesiredDiscoverCmd,
-    state_discover_cmd::StateDiscoverCmd, states_current_discover_cmd::StatesCurrentDiscoverCmd,
+    diff_cmd::DiffCmd, ensure_cmd::EnsureCmd, state_discover_cmd::StateDiscoverCmd,
+    states_current_discover_cmd::StatesCurrentDiscoverCmd,
+    states_desired_discover_cmd::StatesDesiredDiscoverCmd,
 };
 
 mod diff_cmd;
 mod ensure_cmd;
-mod state_desired_discover_cmd;
 mod state_discover_cmd;
 mod states_current_discover_cmd;
+mod states_desired_discover_cmd;