+++ CHANGELOG.md
@@ -5,6 +5,6 @@
 ### Added
 
 * `StatesCurrentDiscoverCmd` to discover the current state.
-* `StateDesiredDiscoverCmd` to discover the desired state.
+* `StatesDesiredDiscoverCmd` to discover the desired state.
 * `DiffCmd` to compute the difference between the current and desired states.
 * `EnsureCmd` to transform the current state into the desired state.